Implement HTTP 1.1 Pipelining
There was an issue with the initial implementation where the BufReader read inside Request::from_tcp_stream would block on requests after the first request.

That was because according to the docs:
When the BufReader<R> is dropped, the contents of its buffer will be discarded. Creating multiple instances of a BufReader<R> on the same stream can cause data loss.
And BufReader was dropped after the first request parsing, so I solved it by putting the BufReader outside the function and passing it as a parameter instead of a TcpStream, this was it wasn't dropped.